mock-server/requirements.txt
fastapi==0.111.0
uvicorn==0.29.0
PyJWT==2.8.0
cryptography==42.0.5
httpx==0.27.0